feat(content-manager): move some files
